+++ Source/Parts/Solids/Ellipsoid.py
