+++ Source/Parts/Solids/__init__.py
@@ -1,0 +1,9 @@
+
+from Source.Parts.Solids.Ellipsoid import *
+from Source.Parts.Solids.Cylinder import *
+from Source.Parts.Solids.Sphere import *
+from Source.Parts.Solids.Prism import *
+from Source.Parts.Solids.Torus import *
+from Source.Parts.Solids.Wedge import *
+from Source.Parts.Solids.Cone import *
+from Source.Parts.Solids.Box import *
