+++ Source/Parts/__init__.py
@@ -1,0 +1,5 @@
+
+from Source.Parts.Object import *
+from Source.Parts.Solid import *
+from Source.Parts.Part import *
+from Source.Parts.Part import *
